Rename chat_model.base to chat_model.message
