+++ src/magentic/chat_model/message.py
