+++ src/magentic/chat_model/openai_chat_model.py
@@ -5,16 +5,16 @@ from typing import Any, Literal, TypeVar, cast
 import openai
 from pydantic import BaseModel, ValidationError
 
-from magentic.chat_model.base import (
-    AssistantMessage,
-    FunctionResultMessage,
-    Message,
-    UserMessage,
-)
 from magentic.chat_model.function_schemas import (
     BaseFunctionSchema,
     FunctionCallFunctionSchema,
     function_schema_for_type,
+)
+from magentic.chat_model.message import (
+    AssistantMessage,
+    FunctionResultMessage,
+    Message,
+    UserMessage,
 )
 from magentic.function_call import FunctionCall
 from magentic.settings import get_settings
